Consolidating internal expression extensions
